+++ lib/extensions/dart_data_type_parsers/int_extensions.dart
@@ -1,0 +1,3 @@
+extension IntExtensions on int {
+  String toStringExt() => '$this';
+}